# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Hello Peer是一款基于兴趣的社交应用。",
        "title": "Hello Peer API",
        "termsOfService": "API文档仅用于Hello Peer研发使用。",
        "contact": {
            "name": "Frank Sheng",
            "email": "726569998@qq.com"
        },
        "version": "0.1"
    },
    "host": "https://hellopeer.net",
    "basePath": "/api/v0.1",
    "paths": {
...
        "/settings/getRecommendSetting": {
            "get": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "description": "可通过用户ID获取用户推荐设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "获取用户推荐设置",
                "operationId": "get_recommend_settings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.RecommendSetting"
                        }
                    }
                }
            }
        },
        "/settings/getSquareSetting": {
            "get": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "description": "可通过用户ID获取用户广场设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "获取用户广场设置",
                "operationId": "get_square_settings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.SquareSettings"
                        }
                    }
                }
            }
        },
        "/settings/setRecommendSetting": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "description": "可设置用户广场设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "设置用户广场设置",
                "operationId": "set_recommend_settings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "设置",
                        "name": "settings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RecommendSetting"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": ""
                    }
                }
            }
        },
        "/settings/setSquareSetting": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "description": "可设置用户广场设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "设置用户广场设置",
                "operationId": "set_square_settings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "设置",
                        "name": "settings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SquareSettings"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": ""
                    }
                }
            }
        },
...
        "/user/pet/deletePetImage": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "description": "删除宠物图片",
                "tags": [
                    "Pet"
                ],
                "summary": "删除宠物图片",
                "operationId": "delete_pet_image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "文件名",
                        "name": "filename",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": ""
                    }
                }
            }
        },
...
        "/user/pet/initPet": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "description": "用户可初始化宠物",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pet"
                ],
                "summary": "初始化宠物",
                "operationId": "init_pet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": ""
                    }
                }
            }
        },
...
        "/user/pet/upload/setPetImage": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "description": "用户可上传宠物图片",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pet"
                ],
                "summary": "设置宠物图片",
                "operationId": "set_pet_image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "宠物图片",
                        "name": "content",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.RecommendSetting": {
            "type": "object",
            "properties": {
                "age_Max": {
                    "type": "integer",
                    "example": 40
                },
                "age_Min": {
                    "type": "integer",
                    "example": 22
                },
                "gender": {
                    "type": "string",
                    "example": "女生"
                },
                "hometown": {
                    "type": "string",
                    "example": "同省优先"
                },
                "location": {
                    "type": "string",
                    "example": "只要同城"
                },
                "pet_Lover": {
                    "type": "string",
                    "example": "喜欢就行"
                },
                "tags": {
                    "type": "string",
                    "example": "不限"
                }
            }
        },
...
        "models.SquareSettings": {
            "type": "object",
            "properties": {
                "gender": {
                    "type": "string",
                    "example": "女生"
                },
                "location": {
                    "type": "string",
                    "example": "不限"
                }
            }
        },
...
    }
}